app.py
import streamlit as st
from PIL import image
import torch
from models.encoder import EncoderCNN
from models.decoder import DecoderTransformer
from inference.caption_generator import generate_caption
from utils.visualization import visualize_attention