requirements.txt
sentence-transformers
pandas
numpy
scikit-learn
pytorch
transformers
tensorflow
opencv
python-dotenv
streamlit